switch around ipc to allow for more complex requests
Envoy no longer automatically starts an agent upon connecting. Instead
it waits for a request to be made. This allows for a cleaner
implementation and more information to be sent along with requests.

Add the option to allow to 'defer' a start.

If a connection is marked 'deferred' envoy will not switch the status
of an agent from ENVOY_STARTED to ENVOY_RUNNING. Since envoy depends on
ENVOY_STARTED to import keys automagically, deferred starts allow us to
avoid complications between pam and envoy-exec.